% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_map.R
\name{add_to_map}
\alias{add_to_map}
\title{Add spatial data to map}
\usage{
add_to_map(map, x, id, col, popup)
}
\arguments{
\item{map}{\code{leaflet} map.}

\item{x}{\code{sf} spatial data.}

\item{id}{\code{character} vector of identifiers for each geometry.}

\item{col}{\code{character} vector of colors.}

\item{popup}{\code{data.frame} with popup data.}
}
\value{
An updated \code{\link[leaflet:leaflet]{leaflet::leaflet()}} map object.
}
\description{
Add spatial data to map
}
