% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{convert_to_html_id}
\alias{convert_to_html_id}
\title{Convert to identifier}
\usage{
convert_to_html_id(x)
}
\arguments{
\item{x}{\code{character} vector containing identifiers.}
}
\value{
A \code{character} value.
}
\description{
Convert a name to a HTML identifier.
}
