% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{convert_to_id}
\alias{convert_to_id}
\title{Convert to identifier}
\usage{
convert_to_id(x)
}
\arguments{
\item{x}{\code{character} value containing a name.}
}
\value{
A \code{character} value.
}
\description{
Convert a name to an identifier.
}
