% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dataModal.R
\name{dataModal}
\alias{dataModal}
\title{Data modal}
\usage{
dataModal(action_ids, action_descriptions, parameters)
}
\arguments{
\item{action_ids}{\code{character} identifiers for actions.}

\item{action_descriptions}{\code{character} descriptions of actions.}

\item{parameters}{\code{list} object with parameters for customizing appearance.}
}
\description{
Constructs a modal for viewing and updating data.
}
\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("dataModal")
}
}

}
