% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dataModalPane.R
\name{dataModalPane}
\alias{dataModalPane}
\title{Data modal pane}
\usage{
dataModalPane(action_ids, action_descriptions, parameters)
}
\arguments{
\item{action_ids}{\code{character} identifiers for actions.}

\item{action_descriptions}{\code{character} descriptions of actions.}

\item{parameters}{\code{list} object with parameters for customizing appearance.}
}
\description{
Constructs a pane for the data modal.
}
\examples{
#TODO

}
