% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dataModalProjectTabset.R
\name{dataModalProjectTab}
\alias{dataModalProjectTab}
\title{Data modal project tabset}
\usage{
dataModalProjectTab(action_ids, action_descriptions, parameters)
}
\arguments{
\item{action_ids}{\code{character} identifiers for actions.}

\item{action_descriptions}{\code{character} descriptions of actions.}

\item{parameters}{\code{list} object with parameters for customizing appearance.}
}
\description{
Constructs a tabset for updating and viewing the project data used to
generate new solutions, and also viewing the data used to generate a
new solution.
}
\examples{
#TODO
}
