% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dataModalResultsTabset.R
\name{dataModalResultsTab}
\alias{dataModalResultsTab}
\title{Data modal results tabset}
\usage{
dataModalResultsTab(action_ids, action_descriptions, parameters)
}
\arguments{
\item{action_ids}{\code{character} identifiers for actions.}

\item{action_descriptions}{\code{character} descriptions of actions.}

\item{parameters}{\code{list} object with parameters for customizing appearance.}
}
\description{
Constructs a tabset for viewing the results of a solution.
}
\examples{
#TODO
}
