% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_defaults.R
\name{default_color_matrix}
\alias{default_color_matrix}
\title{Default color matrix}
\usage{
default_color_matrix(n, parameters)
}
\arguments{
\item{n}{\code{intger} number of sites.}

\item{parameters}{\code{list} object.}
}
\value{
\code{character} \code{matrix} object.
}
\description{
Create a default matrix for specifying colors.
}
