% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_default_colors.R
\name{default_colors}
\alias{default_colors}
\title{Default colors}
\usage{
default_colors(x)
}
\arguments{
\item{x}{\code{character} vector of elements that require colors.}
}
\value{
\code{character} vector of colors (one for each element in \code{x}).
}
\description{
Generate a set of colors.
}
