% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_defaults.R
\name{default_tabular_data}
\alias{default_tabular_data}
\title{Default tabular data}
\usage{
default_tabular_data()
}
\value{
A \code{data.frame} object.
}
\description{
Create a place-holder table to display before any results have been
generated.
}
