% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_errorModal.R
\name{errorModal}
\alias{errorModal}
\title{Solution error modal}
\usage{
errorModal(id, trigger)
}
\arguments{
\item{id}{\code{character} identifier for the modal.}

\item{trigger}{\code{character} identifier for the modal trigger.
See \link[shinyBS:bsModal]{shinyBS::bsModal} for further details.}
}
\value{
A \code{shiny.tag} object with the modal
}
\description{
Constructs a modal for displaying error information.
}
