% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{format_error_data}
\alias{format_error_data}
\title{Format error data}
\usage{
format_error_data(site_names, feature_names, action_names, problem, parameters)
}
\arguments{
\item{site_names}{\code{character} names of sites.}

\item{feature_names}{\code{character} names of features.}

\item{action_names}{\code{character} names of actions.}

\item{problem}{\code{ConservationProblem} object.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{list} object contain \code{data.frame} objects summarizing the results.
}
\description{
Format results for a prioritization problem that threw an error.
}
