% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{format_locked_data}
\alias{format_locked_data}
\title{Format locked data}
\usage{
format_locked_data(
  site_names,
  feature_names,
  action_names,
  site_status_data,
  parameters
)
}
\arguments{
\item{site_names}{\code{character} names of sites.}

\item{feature_names}{\code{character} names of features.}

\item{action_names}{\code{character} names of actions.}

\item{site_status_data}{\code{data.frame} with site status data.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with locked data for prioritizations.
}
\description{
Format locked data
}
