% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{format_pu_data}
\alias{format_pu_data}
\title{Format planning unit data}
\usage{
format_pu_data(
  site_names,
  feature_names,
  action_names,
  site_data,
  action_expectation_data,
  parameters
)
}
\arguments{
\item{site_names}{\code{character} names of sites.}

\item{feature_names}{\code{character} names of features.}

\item{action_names}{\code{character} names of actions.}

\item{site_data}{\code{data.frame} with site data.}

\item{action_expectation_data}{\code{list} of \code{data.frame} objects with
the expected amount of each feature given each action.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with planning unit data for prioritizations.
}
\description{
Format planning unit data
}
