% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{format_results_data}
\alias{format_results_data}
\title{Format results data}
\usage{
format_results_data(
  site_names,
  feature_names,
  action_names,
  pu_data,
  zone_data,
  target_data,
  solution,
  parameters
)
}
\arguments{
\item{site_names}{\code{character} names of sites.}

\item{feature_names}{\code{character} names of features.}

\item{action_names}{\code{character} names of actions.}

\item{pu_data}{\code{data.frame} containing planning unit data used to
create prioritization.}

\item{zone_data}{\code{data.frame} containing zone data used to
create prioritization.}

\item{target_data}{\code{data.frame} containing target data used to
create prioritization.}

\item{solution}{\code{data.frame} containing solution data from
running the prioritization.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{list} object contain \code{data.frame} objects summarizing the results.
}
\description{
Format the results from a prioritization for the application.
}
