% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{format_target_data}
\alias{format_target_data}
\title{Format target data}
\usage{
format_target_data(
  site_names,
  feature_names,
  action_names,
  feature_data,
  parameters
)
}
\arguments{
\item{site_names}{\code{character} names of sites.}

\item{feature_names}{\code{character} names of features.}

\item{action_names}{\code{character} names of actions.}

\item{feature_data}{\code{data.frame} with feature data.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with target data for prioritizations.
}
\description{
Format target data
}
