% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_data.R
\name{format_zone_data}
\alias{format_zone_data}
\title{Format zone data}
\usage{
format_zone_data(site_names, feature_names, action_names, parameters)
}
\arguments{
\item{site_names}{\code{character} names of sites.}

\item{feature_names}{\code{character} names of features.}

\item{action_names}{\code{character} names of actions.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with zone data for prioritizations.
}
\description{
Format zone data
}
