% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_verification.R
\name{has_unique_values}
\alias{has_unique_values}
\title{Has duplicates?}
\usage{
has_unique_values(x)
}
\arguments{
\item{x}{\code{character} \code{vector} object.}
}
\value{
\code{logical} value.
}
\description{
Check if a vector has duplicates?
}
