% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_has_valid_spreadsheet_data.R
\name{has_valid_spreadsheet_data}
\alias{has_valid_spreadsheet_data}
\title{Has valid spreadsheet data?}
\usage{
has_valid_spreadsheet_data(x)
}
\arguments{
\item{x}{\code{list} object.}
}
\value{
\code{logical} indicating valid data.
}
\description{
Check if a spreadsheet has valid data in it.
}
