% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{has_valid_values}
\alias{has_valid_values}
\title{Has valid values?}
\usage{
has_valid_values(x)
}
\arguments{
\item{x}{\code{character} \code{vector} object.}
}
\value{
\code{logical} value.
}
\description{
Check if a vector has valid values?
}
