% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_data.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(x)
}
\arguments{
\item{x}{\code{list} with project data.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Import project data into the app.
}
