% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_initial_ui.R
\name{initial_data_ui}
\alias{initial_data_ui}
\title{Initial data user interface}
\usage{
initial_data_ui()
}
\value{
A \code{shiny.tag} object.
}
\description{
Create a user interface component as a place-holder to display data before
any data are uploaded to the application.
}
