% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_initial_ui.R
\name{initial_map_ui}
\alias{initial_map_ui}
\title{Initial map user interface}
\usage{
initial_map_ui()
}
\value{
A \code{shiny.tag} object.
}
\description{
Create a user interface component as a place-holder to display
a map before any data are uploaded to the application.
}
