% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_initial_ui.R
\name{initial_results_ui}
\alias{initial_results_ui}
\title{Initial result user interface}
\usage{
initial_results_ui()
}
\value{
A \code{shiny.tag} object.
}
\description{
Create a user interface component as a place-holder to display results
before any data are uploaded to the application.
}
