% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_initial_ui.R
\name{initial_sidebar_ui}
\alias{initial_sidebar_ui}
\title{Initial sidebar user interface}
\usage{
initial_sidebar_ui()
}
\value{
A \code{shiny.tag} object.
}
\description{
Create a user interface component for uploading data to the the application.
}
