% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_map.R
\name{initialize_map}
\alias{initialize_map}
\title{Initialize map}
\usage{
initialize_map(site_bbox, site_spatial_data, parameters)
}
\arguments{
\item{site_bbox}{\code{bbox} bounding box for site data.}

\item{site_spatial_data}{\code{sf} spatial data for sites.}

\item{parameters}{\code{list} parameters for interpreting the data.
Defaults to configuration parameters bundled with the application.}
}
\value{
A \code{\link[leaflet:leaflet]{leaflet::leaflet()}} map object.
}
\description{
Create a (\pkg{leaflet}) map. This map is shown by default
after data are loaded into the application, before any solutions are
generated.
}
