% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_is_valid_shapefile.R
\name{is_valid_shapefile}
\alias{is_valid_shapefile}
\title{Is valid shapefile?}
\usage{
is_valid_shapefile(x)
}
\arguments{
\item{x}{\code{character} file path.}
}
\value{
\code{logical} indicating validity.
}
\description{
Determine if a file is a valid shapefile?
}
