% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{isFALSE}
\alias{isFALSE}
\title{Is false?}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{an object.}
}
\value{
\code{logical} indicating if object is identical to \code{FALSE}.
}
\description{
Determine if an object is identical to \code{FALSE}.
}
