% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mapSidebarPane.R
\name{mapSidebarPane}
\alias{mapSidebarPane}
\title{Map sidebar pane}
\usage{
mapSidebarPane(
  id,
  dataset_id = paste0(id, "_dataset"),
  layer_id = paste0(id, "_layer")
)
}
\arguments{
\item{id}{\code{character} identifier for the sidebar pane.}
}
\description{
Constructs a sidebar pane for showing data on the map.
}
\examples{
# TODO

}
