% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Solution.R
\name{new_solution}
\alias{new_solution}
\title{New solution}
\usage{
new_solution(
  project,
  name,
  settings,
  summary_results,
  site_results,
  feature_results,
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{project}{\link{Project} object.}

\item{name}{\code{character} value with name for solution.}

\item{settings}{\code{list} of \link{Parameter} objects.}

\item{summary_results}{\code{data.frame} containing the summary results.}

\item{site_results}{\code{data.frame} containing the site results.}

\item{feature_results}{\code{data.frame} containing the feature results.}

\item{id}{\code{character} identifier for object.}
}
\value{
A \link{Solution} object.
}
\description{
Create a new \link{Solution} object.
}
\examples{
#TODO
}
