% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize_data.R
\name{num_whole_digits}
\alias{num_whole_digits}
\title{Count Whole Number Digits}
\usage{
num_whole_digits(x)
}
\arguments{
\item{x}{A \code{numeric} value or vector.}
}
\value{
An \code{integer} representing the number of whole number digits.
}
\description{
Determines the number of whole number digits in a numeric value.
}
\details{
Calculates the number of digits before the decimal point
in the absolute value of \code{x}. It truncates any decimal places and
ensures that scientific notation is not used when counting digits.
}
\examples{
num_whole_digits(12345)   # Returns 5
num_whole_digits(0.1234)  # Returns 1
num_whole_digits(1e6)     # Returns 7

}
