% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Parameter.R
\name{Parameter}
\alias{Parameter}
\title{Parameter class}
\description{
Parameter class

Parameter class
}
\details{
Definition for the Parameter class.
}
\seealso{
\code{\link[=new_parameter]{new_parameter()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{value}}{\code{numeric} initial value value.}

\item{\code{min_value}}{\code{numeric} minimum value.}

\item{\code{max_value}}{\code{numeric} maximum value.}

\item{\code{step_value}}{\code{numeric} step value.}

\item{\code{hide}}{\code{logical} value.}

\item{\code{units}}{\code{character} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Parameter-new}{\code{Parameter$new()}}
\item \href{#method-Parameter-print}{\code{Parameter$print()}}
\item \href{#method-Parameter-repr}{\code{Parameter$repr()}}
\item \href{#method-Parameter-get_status}{\code{Parameter$get_status()}}
\item \href{#method-Parameter-get_value}{\code{Parameter$get_value()}}
\item \href{#method-Parameter-get_setting}{\code{Parameter$get_setting()}}
\item \href{#method-Parameter-set_status}{\code{Parameter$set_status()}}
\item \href{#method-Parameter-set_value}{\code{Parameter$set_value()}}
\item \href{#method-Parameter-set_setting}{\code{Parameter$set_setting()}}
\item \href{#method-Parameter-get_widget_data}{\code{Parameter$get_widget_data()}}
\item \href{#method-Parameter-export}{\code{Parameter$export()}}
\item \href{#method-Parameter-get_results_data}{\code{Parameter$get_results_data()}}
\item \href{#method-Parameter-clone}{\code{Parameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-new}{}}}
\subsection{Method \code{new()}}{
Create a new Parameter object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$new(
  id,
  name,
  status,
  value,
  min_value,
  max_value,
  step_value,
  hide,
  units
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{name}}{\code{character} value.}

\item{\code{status}}{\code{logical} value.}

\item{\code{value}}{\code{numeric} initial value.}

\item{\code{min_value}}{\code{numeric} minimum value.}

\item{\code{max_value}}{\code{numeric} maximum value.}

\item{\code{step_value}}{\code{numeric} step value.}

\item{\code{hide}}{\code{logical} value.}

\item{\code{units}}{\code{character} value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Parameter object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-print"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-repr"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$repr(start = "[", end = "]")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\code{character} symbol used to start the setting list.
Defaults to \code{"["}.}

\item{\code{end}}{\code{character} symbol used to start the setting list.
Defaults to \code{"]"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-get_value}{}}}
\subsection{Method \code{get_value()}}{
Get value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-get_setting"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-get_setting}{}}}
\subsection{Method \code{get_setting()}}{
Get setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_setting(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"status"}, or \code{"value"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-set_status"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-set_status}{}}}
\subsection{Method \code{set_status()}}{
Set status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$set_status(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{logical} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-set_value}{}}}
\subsection{Method \code{set_value()}}{
Set value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$set_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{\code{numeric} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-set_setting"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-set_setting}{}}}
\subsection{Method \code{set_setting()}}{
Set setting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$set_setting(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} setting name.
Available options are \code{"status"}, or \code{"value"}.}

\item{\code{value}}{\code{ANY} new value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-get_widget_data"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-get_widget_data}{}}}
\subsection{Method \code{get_widget_data()}}{
Get data for displaying the parameter in a widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_widget_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-export"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-export}{}}}
\subsection{Method \code{export()}}{
Export settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-get_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-get_results_data}{}}}
\subsection{Method \code{get_results_data()}}{
Get results data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$get_results_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
