% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{parse_action_expectation_data}
\alias{parse_action_expectation_data}
\title{Parse action expectation data}
\usage{
parse_action_expectation_data(x)
}
\arguments{
\item{x}{\code{rhandsontable} object.}
}
\value{
x \code{data.frame} object.
}
\description{
Extract action expectation data from \pkg{rhandsontable}.
}
