% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{parse_feature_data}
\alias{parse_feature_data}
\title{Parse feature data}
\usage{
parse_feature_data(x)
}
\arguments{
\item{x}{\code{rhandsontable} object.}
}
\value{
x \code{data.frame} object.
}
\description{
Extract feature data from \pkg{rhandsontable}.
}
