% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{parse_site_data}
\alias{parse_site_data}
\title{Parse site data}
\usage{
parse_site_data(x)
}
\arguments{
\item{x}{\code{rhandsontable} object.}
}
\value{
x \code{data.frame} object.
}
\description{
Extract site data from \pkg{rhandsontable}.
}
