% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{paste_vector}
\alias{paste_vector}
\title{Paste vector}
\usage{
paste_vector(x)
}
\arguments{
\item{x}{\code{character} vector.}
}
\value{
\code{character} value.
}
\description{
Paste a \code{character} vector of together for displaying messages.
}
