% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_for_shapefile_import.R
\name{prepare_for_shapefile_import}
\alias{prepare_for_shapefile_import}
\title{Prepare files for shapefile import}
\usage{
prepare_for_shapefile_import(x)
}
\arguments{
\item{x}{\code{character} file paths.}
}
\value{
\code{character} file paths for renamed files.
}
\description{
This function prepares a set of files so that they can be imported as
a shapefile.
}
\details{
Specifically, the files are all assigned the same file name
(i.e. base name without the file extension) so that they form the
constituent files of a shapefile.
}
