% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_primary_sidebar_ui.R
\name{primary_sidebar_ui}
\alias{primary_sidebar_ui}
\title{Sidebar user interface}
\usage{
primary_sidebar_ui(site_names, feature_names, action_names, parameters)
}
\arguments{
\item{site_names}{\code{character} names of sites.}

\item{feature_names}{\code{character} names of features.}

\item{action_names}{\code{character} names of actions.}

\item{parameters}{\code{list} object with parameters for customizing appearance.}
}
\value{
A \code{shiny.tag} object.
}
\description{
Create a sidebar user interface for generating prioritizations.
}
