% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prioritization_with_budget.R
\name{prioritization_with_budget}
\alias{prioritization_with_budget}
\title{Prioritization without budget}
\usage{
prioritization_with_budget(
  site_ids,
  feature_ids,
  action_ids,
  pu_data,
  status_data,
  zone_data,
  goal_data,
  weight_data,
  locked_data,
  budget,
  max_budget,
  parameters,
  gap = 0,
  verbose = TRUE,
  time_limit = .Machine$integer.max
)
}
\arguments{
\item{site_ids}{\code{character} site identifiers.}

\item{feature_ids}{\code{character} feature identifiers.}

\item{action_ids}{\code{character} action identifiers.}

\item{pu_data}{\code{list} containing raw planning unit data as \code{data.frame} and
cost normalized planning unit data as either \code{data.frame} or \code{NULL}.}

\item{status_data}{\code{dat.frame} containing status data.}

\item{zone_data}{\code{dat.frame} containing zone data.}

\item{goal_data}{\code{dat.frame} containing goal data.}

\item{weight_data}{\code{dat.frame} containing weight data.}

\item{locked_data}{\code{dat.frame} containing locked out data.}

\item{budget}{\code{numeric} budget.}

\item{parameters}{\code{list} object with configuration parameters.}

\item{gap}{\code{numeric} optimality gap. Defaults to 0.}

\item{verbose}{\code{logical} display information during optimization?
Defaults to \code{TRUE}.}

\item{time_limit}{\code{integer} Maximum amount of time (seconds) to spend
during optimization.
Defaults to the maximum integer value.}
}
\value{
\code{list} object containing results of prioritization.
}
\description{
Generate a prioritization without a budget.
}
