% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_builtin_project.R
\name{read_builtin_project}
\alias{read_builtin_project}
\title{Read builtin project}
\usage{
read_builtin_project(path, parameters)
}
\arguments{
\item{path}{\code{character} file path for the project YAML metadata file.}

\item{parameters}{\code{list} object to interpret spreadsheet data.}
}
\value{
A \code{list} containing the following elements:
\describe{
\item{name}{A \code{character} value indicating the name.}
\item{author_name}{A \code{character} indicating the author name for the data.}
\item{author_email}{A \code{character} indicating the contact email for the data.}
\item{project}{A \link{Project} object.}
}
}
\description{
Read a project from disk.
}
\examples{
# specify file path
f1 <- system.file(
  "extdata", "projects", "simulated-data", "simulated-data.yaml",
  package = "whattodo"
)

# read project
x <- read_builtin_project(f1, whatdataio::read_data_configuration())

# print project
print(x)
}
