% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_render_data.R
\name{render_action_expectation_data}
\alias{render_action_expectation_data}
\title{Render action expectation data}
\usage{
render_action_expectation_data(x)
}
\arguments{
\item{x}{\code{data.frame} object.}
}
\description{
Display action expectation data using \pkg{rhandsontable}.
}
