% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_render_data.R
\name{render_feature_data}
\alias{render_feature_data}
\title{Render feature data}
\usage{
render_feature_data(x)
}
\arguments{
\item{x}{\code{data.frame} object.}
}
\description{
Display feature data using \pkg{rhandsontable}.
}
