% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_render_data.R
\name{render_site_data}
\alias{render_site_data}
\title{Render site data}
\usage{
render_site_data(x)
}
\arguments{
\item{x}{\code{data.frame} object.}
}
\value{
\code{rhandsontable} object.
}
\description{
Display site data using \pkg{rhandsontable}.
}
