% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_render_data.R
\name{render_site_status_data}
\alias{render_site_status_data}
\title{Render status data}
\usage{
render_site_status_data(x, parameters)
}
\arguments{
\item{x}{\code{data.frame} object.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
A \code{rhandsontable} object.
}
\description{
Display site status data using \pkg{rhandsontable}.
}
