% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_render_data.R
\name{render_summary_results_data}
\alias{render_summary_results_data}
\title{Render summary results data}
\usage{
render_summary_results_data(x, parameters)
}
\arguments{
\item{x}{\code{data.frame} object.}

\item{parameters}{\code{list} object.}
}
\description{
Display summary results data using \pkg{rhandontable}.
}
