% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_map.R
\name{reset_map}
\alias{reset_map}
\title{Reset map}
\usage{
reset_map(map, site_spatial_data, parameters, legend_title = NULL)
}
\arguments{
\item{map}{\code{leaflet} map object from \code{leafletProxy}.}

\item{site_spatial_data}{\code{sf} spatial data for sites.}

\item{parameters}{\code{list} parameters for interpreting the data.
Defaults to configuration parameters bundled with the application.}

\item{legend_title}{\code{character} title of legend.}
}
\value{
A \code{\link[leaflet:leaflet]{leaflet::leaflet()}} map object.
}
\description{
Reset map
}
