% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Solution.R
\name{Solution}
\alias{Solution}
\title{Solution class}
\description{
Solution class

Solution class
}
\details{
Definition for the Solution class.
}
\seealso{
\code{\link[=new_project]{new_project()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{name}}{\code{character} name.}

\item{\code{project}}{\link{Project} object.}

\item{\code{settings}}{\code{list} of \link{Parameter} objects.}

\item{\code{summary_results}}{\code{data.frame} object.}

\item{\code{site_results}}{\code{data.frame} object.}

\item{\code{feature_results}}{\code{data.frame} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Solution-new}{\code{Solution$new()}}
\item \href{#method-Solution-print}{\code{Solution$print()}}
\item \href{#method-Solution-repr}{\code{Solution$repr()}}
\item \href{#method-Solution-get_site_ids}{\code{Solution$get_site_ids()}}
\item \href{#method-Solution-get_feature_ids}{\code{Solution$get_feature_ids()}}
\item \href{#method-Solution-get_action_ids}{\code{Solution$get_action_ids()}}
\item \href{#method-Solution-get_map_layers}{\code{Solution$get_map_layers()}}
\item \href{#method-Solution-get_bbox}{\code{Solution$get_bbox()}}
\item \href{#method-Solution-render_on_map}{\code{Solution$render_on_map()}}
\item \href{#method-Solution-get_solution_results_data}{\code{Solution$get_solution_results_data()}}
\item \href{#method-Solution-render_site_data}{\code{Solution$render_site_data()}}
\item \href{#method-Solution-render_feature_data}{\code{Solution$render_feature_data()}}
\item \href{#method-Solution-render_feasibility_data}{\code{Solution$render_feasibility_data()}}
\item \href{#method-Solution-render_consequence_data}{\code{Solution$render_consequence_data()}}
\item \href{#method-Solution-render_summary_results}{\code{Solution$render_summary_results()}}
\item \href{#method-Solution-render_site_results}{\code{Solution$render_site_results()}}
\item \href{#method-Solution-render_feature_results}{\code{Solution$render_feature_results()}}
\item \href{#method-Solution-write}{\code{Solution$write()}}
\item \href{#method-Solution-clone}{\code{Solution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-new"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-new}{}}}
\subsection{Method \code{new()}}{
Create a Solution object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$new(
  id,
  name,
  project,
  settings,
  summary_results,
  site_results,
  feature_results
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{name}}{\code{character} value.}

\item{\code{project}}{\code{character} identifier.}

\item{\code{settings}}{\code{list} of \link{Parameter} objects.}

\item{\code{summary_results}}{\code{data.frame} object.}

\item{\code{site_results}}{\code{data.frame} object.}

\item{\code{feature_results}}{\code{data.frame} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Solution object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-print"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-repr"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_site_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_site_ids}{}}}
\subsection{Method \code{get_site_ids()}}{
Get site identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_site_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_feature_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_feature_ids}{}}}
\subsection{Method \code{get_feature_ids()}}{
Get feature identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_feature_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_action_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_action_ids}{}}}
\subsection{Method \code{get_action_ids()}}{
Get action identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_action_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_map_layers"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_map_layers}{}}}
\subsection{Method \code{get_map_layers()}}{
Get layer names for rendering data on map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_map_layers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_bbox"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_bbox}{}}}
\subsection{Method \code{get_bbox()}}{
Get the bounding box.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_bbox(expand = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expand}}{\code{FALSE} should the bounding box be expanded by 10\%?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} object with \code{"xmin"}, \code{"xmax"}, \code{"ymin"}, and \code{"ymax"}
elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_on_map(map, data = "location", group = "sites")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{data}}{\code{character} name of dataset to show.
Argument must be a valid layer name (see \code{self$get_map_layers()}).}

\item{\code{group}}{\code{character} group name. Defaults to \code{"sites"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} map.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-get_solution_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-get_solution_results_data}{}}}
\subsection{Method \code{get_solution_results_data()}}{
Get data for rendering widget to display results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_solution_results_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_site_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_site_data}{}}}
\subsection{Method \code{render_site_data()}}{
Render site data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_site_data(height = "100\%", width = "100\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}

\item{\code{width}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[rhandsontable:rhandsontable]{rhandsontable::rhandsontable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_feature_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_feature_data}{}}}
\subsection{Method \code{render_feature_data()}}{
Render feature data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_feature_data(height = "100\%", width = "100\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}

\item{\code{width}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[rhandsontable:rhandsontable]{rhandsontable::rhandsontable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_feasibility_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_feasibility_data}{}}}
\subsection{Method \code{render_feasibility_data()}}{
Render feasibility data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_feasibility_data(height = "100\%", width = "100\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}

\item{\code{width}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[rhandsontable:rhandsontable]{rhandsontable::rhandsontable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_consequence_data"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_consequence_data}{}}}
\subsection{Method \code{render_consequence_data()}}{
Render consequence data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_consequence_data(action_id, height = "100\%", width = "100\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}

\item{\code{height}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}

\item{\code{width}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[rhandsontable:rhandsontable]{rhandsontable::rhandsontable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_summary_results"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_summary_results}{}}}
\subsection{Method \code{render_summary_results()}}{
Display summary results data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_summary_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[DT:datatable]{DT::datatable()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_site_results"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_site_results}{}}}
\subsection{Method \code{render_site_results()}}{
Display site results data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_site_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[DT:datatable]{DT::datatable()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-render_feature_results"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-render_feature_results}{}}}
\subsection{Method \code{render_feature_results()}}{
Display feature results data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_feature_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[DT:datatable]{DT::datatable()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-write"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-write}{}}}
\subsection{Method \code{write()}}{
Write the data to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$write(workbook_path, geometry_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workbook_path}}{\code{character} file path.}

\item{\code{geometry_path}}{\code{character} file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Solution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Solution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
