% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solution_action_consequence.R
\name{solution_action_consequence}
\alias{solution_action_consequence}
\title{Calculate action consequence data for solution}
\usage{
solution_action_consequence(feature_ids, action_ids, pu_data, solution_data)
}
\arguments{
\item{feature_ids}{\code{character} vector of identifiers for each feature.}

\item{action_ids}{\code{character} vector of identifiers for each action.}

\item{pu_data}{\code{data.frame} containing planning unit data.}

\item{solution_data}{\code{data.frame} object containing the solution.}
}
\value{
A \code{matrix} containing the expected amount of each
feature resulting from each action in the solution.
Each row is a different feature, each column is a different action.
}
\description{
This function calculates the consequences for each feature
within a solution. Specifically, it shows how each action contributes
to the overall amount of each feature given the solution.
}
