% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solution_site_consequence.R
\name{solution_site_consequence}
\alias{solution_site_consequence}
\title{Calculate site consequence data for solution}
\usage{
solution_site_consequence(feature_ids, action_ids, pu_data, solution_data)
}
\arguments{
\item{feature_ids}{\code{character} vector of identifiers for each feature.}

\item{action_ids}{\code{character} vector of identifiers for each action.}

\item{pu_data}{\code{data.frame} containing planning unit data.}

\item{solution_data}{\code{data.frame} object containing the solution.}
}
\value{
A \code{matrix} object containing the expected
amount of each feature within each site.
Each row is a different planning unit, each column is a different feature.
}
\description{
This function calculates the consequences for each site
within a solution. Specifically, it shows how each site contributes
to the overall amount of each feature given the solution.
}
