% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_solutionModal.R
\name{solutionModal}
\alias{solutionModal}
\title{Solution modal}
\usage{
solutionModal(id, trigger)
}
\arguments{
\item{id}{\code{character} identifier for the modal.}

\item{trigger}{\code{character} identifier for the widget to trigger the modal.}
}
\description{
Constructs a modal for viewing solution data and results.
}
\examples{
\dontrun{
# run Shiny app to demo the sidebar pane
if (interactive()) {
  runExample("solutionModal")
}
}

}
