% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sf.R
\name{st_geometry_data_type}
\alias{st_geometry_data_type}
\title{Geometry data type}
\usage{
st_geometry_data_type(x)
}
\arguments{
\item{x}{\code{sf} object.}
}
\value{
\code{character} description of data type.
}
\description{
Determine type of geometry in \code{sf} object.
}
