% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_map.R
\name{update_map}
\alias{update_map}
\title{Update map}
\usage{
update_map(
  map,
  feature_names,
  site_spatial_data,
  site_results_data,
  action_expectation_data,
  action_colors,
  parameters,
  legend_title = NULL
)
}
\arguments{
\item{map}{\code{leaflet} map object from \code{leafletProxy}.}

\item{feature_names}{\code{character} names of features.}

\item{site_spatial_data}{\code{sf} spatial data for sites.}

\item{site_results_data}{\code{data.frame} object with results from
prioritization.}

\item{action_expectation_data}{\code{list} object with action expectation data.}

\item{action_colors}{\code{character} colors to use for rendering prioritizations.}

\item{parameters}{\code{list} parameters for interpreting the data.
Defaults to configuration parameters bundled with the application.}

\item{legend_title}{\code{character} title of legend.}
}
\value{
A \code{\link[leaflet:leaflet]{leaflet::leaflet()}} map object.
}
\description{
Update map
}
