% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dataModal.R
\name{dataModal}
\alias{dataModal}
\title{Data modal}
\usage{
dataModal(id, trigger)
}
\arguments{
\item{id}{\code{character} identifier for the modal.}

\item{trigger}{\code{character} identifier for the widget to trigger the modal.}
}
\description{
Constructs a modal for viewing and editing project data, and examining the
performance of solutions.
}
\examples{
#TODO
}
