% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_defaults.R
\name{default_colors}
\alias{default_colors}
\title{Default colors}
\usage{
default_colors(x, parameters)
}
\arguments{
\item{x}{\code{character} vector of elements that require colors.}

\item{parameters}{\code{list} object.}
}
\value{
\code{character} vector of colors (one for each element in \code{x})
}
\description{
Generate a set of colors.
}
