% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_solution_error.R
\name{format_error_data}
\alias{format_error_data}
\title{Format error data}
\usage{
format_error_data(parameters)
}
\arguments{
\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{list} object containing \code{data.frame} objects with results.
}
\description{
Format results for a prioritization problem that threw an error.
}
