% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_solution_results.R
\name{format_solution_results}
\alias{format_solution_results}
\title{Format solution results}
\usage{
format_solution_results(
  site_ids,
  feature_ids,
  action_ids,
  pu_data,
  zone_data,
  goal_data,
  status_data,
  locked_data,
  solution_data,
  budget,
  parameters
)
}
\arguments{
\item{site_ids}{\code{character} site identifiers.}

\item{feature_ids}{\code{character} feature identifiers.}

\item{action_ids}{\code{character} action identifiers.}

\item{pu_data}{\code{dat.frame} containing planning unit data.}

\item{zone_data}{\code{dat.frame} containing zone data.}

\item{goal_data}{\code{dat.frame} containing goal data.}

\item{status_data}{\code{dat.frame} containing status data.}

\item{locked_data}{\code{dat.frame} containing locked out data.}

\item{solution_data}{\code{data.frame} containing solution data from
running the prioritization.}

\item{budget}{\code{numeric} budget.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{list} object containing \code{data.frame} objects with results.
}
\description{
Format the results from a prioritization for the application.
}
