% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Parameter.R
\name{new_parameter}
\alias{new_parameter}
\title{New parameter}
\usage{
new_parameter(
  name,
  status = TRUE,
  value = 0,
  min_value = 0,
  max_value = 100,
  step_value = 1,
  hide = FALSE,
  units = "",
  id = uuid::UUIDgenerate()
)
}
\arguments{
\item{name}{\code{character} Name for parameter.}

\item{status}{\code{logical} status for parameter.}

\item{value}{\code{numeric} initial value.
Defaults to 0.}

\item{min_value}{\code{numeric} minimum value.
Defaults to 0.}

\item{max_value}{\code{numeric} maximum value.}

\item{step_value}{\code{numeric} step value.
Defaults to 1.}

\item{hide}{\code{logical} value indicating if the slider should be
hidden when the parameter is disabled.
Defaults to \code{FALSE}.}

\item{units}{\code{character} units.
Defaults to an empty \code{character} object.}

\item{id}{\code{character} identifier for object.}
}
\value{
A \link{Parameter} object.
}
\description{
Create a new \link{Parameter} object.
}
\examples{
# create a new parameter
p <- new_parameter(name = "Spatial clustering")

# print object
print(p)
}
