% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prioritization_with_budget.R
\name{prioritization_with_budget}
\alias{prioritization_with_budget}
\title{Prioritization without budget}
\usage{
prioritization_with_budget(
  site_ids,
  feature_ids,
  action_ids,
  pu_data,
  zone_data,
  goal_data,
  weight_data,
  locked_data,
  budget,
  gap = 0,
  parameters
)
}
\arguments{
\item{site_ids}{\code{character} site identifiers.}

\item{feature_ids}{\code{character} feature identifiers.}

\item{action_ids}{\code{character} action identifiers.}

\item{pu_data}{\code{dat.frame} containing planning unit data.}

\item{zone_data}{\code{dat.frame} containing zone data.}

\item{goal_data}{\code{dat.frame} containing goal data.}

\item{weight_data}{\code{dat.frame} containing weight data.}

\item{locked_data}{\code{dat.frame} containing locked out data.}

\item{budget}{\code{numeric} budget.}

\item{gap}{\code{numeric} optimality gap.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{list} object containing results of prioritization.
}
\description{
Generate a prioritization without a budget.
}
