% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prioritization_with_budget.R
\name{prioritization_with_budget}
\alias{prioritization_with_budget}
\title{Prioritization without budget}
\usage{
prioritization_with_budget(
  site_names,
  feature_names,
  action_names,
  site_data,
  site_status_data,
  feature_data,
  action_expectation_data,
  budget,
  weights = TRUE,
  gap = 0,
  parameters
)
}
\arguments{
\item{site_names}{\code{character} names of sites.}

\item{feature_names}{\code{character} names of features.}

\item{action_names}{\code{character} names of actions.}

\item{site_data}{\code{data.frame} with site data.}

\item{site_status_data}{\code{data.frame} with site status data.}

\item{feature_data}{\code{data.frame} with feature data.}

\item{action_expectation_data}{\code{list} of \code{data.frame} objects with
the expected amount of each feature given each action.}

\item{budget}{\code{numeric} budget.}

\item{weights}{\code{logical} use weights?}

\item{gap}{\code{numeric} optimality gap.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{list} object containing results of prioritization.
}
\description{
Generate a prioritization without a budget.
}
