% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Project.R
\name{Project}
\alias{Project}
\title{Project class}
\description{
Project class

Project class
}
\details{
Definition for the Project class.
}
\seealso{
\code{\link[=new_project]{new_project()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{site_ids}}{\code{character} vector}

\item{\code{site_descriptions}}{\code{character} vector}

\item{\code{feature_ids}}{\code{character} vector}

\item{\code{feature_descriptions}}{\code{character} vector}

\item{\code{action_ids}}{\code{character} vector}

\item{\code{action_colors}}{\code{character} vector of colors for each action.}

\item{\code{action_descriptions}}{\code{character} vector}

\item{\code{site_data}}{\code{data.frame} object.}

\item{\code{feasibility_data}}{\code{data.frame} object.}

\item{\code{site_geometry}}{\code{\link[sf:sf]{sf::st_sf()}} object.}

\item{\code{parameters}}{\code{list} of parameters.}

\item{\code{settings}}{\code{list} of \code{Parameter} objects.}

\item{\code{action_expectation_data}}{\code{list} of \code{data.frame} objects.}

\item{\code{site_cost_headers}}{\code{character} vector with names.}

\item{\code{site_status_header}}{\code{character} value.}

\item{\code{action_feasibility_headers}}{\code{character} vector with names.}

\item{\code{feature_goal_header}}{\code{character} value.}

\item{\code{feature_weight_header}}{\code{character} value.}

\item{\code{action_expectation_action_headers}}{\code{character} named vector.}

\item{\code{action_expectation_feature_headers}}{\code{character} named vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Project$new()}}
\item \href{#method-print}{\code{Project$print()}}
\item \href{#method-repr}{\code{Project$repr()}}
\item \href{#method-get_site_ids}{\code{Project$get_site_ids()}}
\item \href{#method-get_feature_ids}{\code{Project$get_feature_ids()}}
\item \href{#method-get_action_ids}{\code{Project$get_action_ids()}}
\item \href{#method-get_site_statuses}{\code{Project$get_site_statuses()}}
\item \href{#method-get_action_costs}{\code{Project$get_action_costs()}}
\item \href{#method-get_action_feasibility}{\code{Project$get_action_feasibility()}}
\item \href{#method-get_action_expectations_for_feature}{\code{Project$get_action_expectations_for_feature()}}
\item \href{#method-get_action_expectation_feature_maxima}{\code{Project$get_action_expectation_feature_maxima()}}
\item \href{#method-get_pu_data}{\code{Project$get_pu_data()}}
\item \href{#method-get_zone_data}{\code{Project$get_zone_data()}}
\item \href{#method-get_goal_data}{\code{Project$get_goal_data()}}
\item \href{#method-get_weight_data}{\code{Project$get_weight_data()}}
\item \href{#method-get_locked_data}{\code{Project$get_locked_data()}}
\item \href{#method-get_bbox}{\code{Project$get_bbox()}}
\item \href{#method-add_to_map}{\code{Project$add_to_map()}}
\item \href{#method-get_solution_settings_data}{\code{Project$get_solution_settings_data()}}
\item \href{#method-get_goals_settings_data}{\code{Project$get_goals_settings_data()}}
\item \href{#method-get_weights_settings_data}{\code{Project$get_weights_settings_data()}}
\item \href{#method-set_feature_goal}{\code{Project$set_feature_goal()}}
\item \href{#method-set_feature_weight}{\code{Project$set_feature_weight()}}
\item \href{#method-set_site_data}{\code{Project$set_site_data()}}
\item \href{#method-set_feature_data}{\code{Project$set_feature_data()}}
\item \href{#method-set_feasibility_data}{\code{Project$set_feasibility_data()}}
\item \href{#method-set_action_expectation_data}{\code{Project$set_action_expectation_data()}}
\item \href{#method-render_site_data}{\code{Project$render_site_data()}}
\item \href{#method-render_feature_data}{\code{Project$render_feature_data()}}
\item \href{#method-render_feasibility_data}{\code{Project$render_feasibility_data()}}
\item \href{#method-render_action_expectation_data}{\code{Project$render_action_expectation_data()}}
\item \href{#method-write}{\code{Project$write()}}
\item \href{#method-clone}{\code{Project$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Project object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$new(
  id,
  site_ids,
  site_descriptions,
  site_geometry,
  feature_ids,
  feature_descriptions,
  action_ids,
  action_descriptions,
  site_data,
  feasibility_data,
  action_expectation_data,
  parameters
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{site_ids}}{\code{character} vector}

\item{\code{site_descriptions}}{\code{character} vector}

\item{\code{site_geometry}}{\code{\link[sf:sf]{sf::st_sf()}} object.}

\item{\code{feature_ids}}{\code{character} vector}

\item{\code{feature_descriptions}}{\code{character} vector}

\item{\code{action_ids}}{\code{character} vector}

\item{\code{action_descriptions}}{\code{character} vector}

\item{\code{site_data}}{\code{data.frame} object.}

\item{\code{feasibility_data}}{\code{data.frame} object.}

\item{\code{action_expectation_data}}{\code{list} of \code{data.frame} objects.}

\item{\code{parameters}}{\code{list} of parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Project object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_site_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_site_ids}{}}}
\subsection{Method \code{get_site_ids()}}{
Get site identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_site_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_ids}{}}}
\subsection{Method \code{get_feature_ids()}}{
Get feature identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_feature_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_ids}{}}}
\subsection{Method \code{get_action_ids()}}{
Get action identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_action_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_site_statuses"></a>}}
\if{latex}{\out{\hypertarget{method-get_site_statuses}{}}}
\subsection{Method \code{get_site_statuses()}}{
Get colors for actions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_site_statuses()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_costs"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_costs}{}}}
\subsection{Method \code{get_action_costs()}}{
Get cost data for implementing an action within each site.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_action_costs(action_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{numeric} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_feasibility"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_feasibility}{}}}
\subsection{Method \code{get_action_feasibility()}}{
Get feasibility data for implementing an action within each site.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_action_feasibility(action_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_expectations_for_feature"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_expectations_for_feature}{}}}
\subsection{Method \code{get_action_expectations_for_feature()}}{
Get action expectation data for features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_action_expectations_for_feature(action_id, feature_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}

\item{\code{feature_id}}{\code{character} identifier for feature.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{numeric} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_expectation_feature_maxima"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_expectation_feature_maxima}{}}}
\subsection{Method \code{get_action_expectation_feature_maxima()}}{
Get the highest amount for each feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_action_expectation_feature_maxima()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pu_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_pu_data}{}}}
\subsection{Method \code{get_pu_data()}}{
Get planning unit data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_pu_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_zone_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_zone_data}{}}}
\subsection{Method \code{get_zone_data()}}{
Get zone data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_zone_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_goal_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_goal_data}{}}}
\subsection{Method \code{get_goal_data()}}{
Get goal data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_goal_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weight_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_weight_data}{}}}
\subsection{Method \code{get_weight_data()}}{
Get weight data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_weight_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_locked_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_locked_data}{}}}
\subsection{Method \code{get_locked_data()}}{
Get locked data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_locked_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bbox"></a>}}
\if{latex}{\out{\hypertarget{method-get_bbox}{}}}
\subsection{Method \code{get_bbox()}}{
Get the bounding box.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_bbox(native = TRUE, expand = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{native}}{\code{logical} indicating if the bounding box should
be in (\code{TRUE}) the native coordinate reference system or (\code{FALSE})
re-projected to longitude/latitude?}

\item{\code{expand}}{\code{FALSE} should the bounding box be expanded by 10\%?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} object with \code{"xmin"}, \code{"xmax"}, \code{"ymin"}, and \code{"ymax"}
elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_to_map"></a>}}
\if{latex}{\out{\hypertarget{method-add_to_map}{}}}
\subsection{Method \code{add_to_map()}}{
Add to map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$add_to_map(map, data = "location", group = "sites")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{data}}{\code{character} name of dataset to show.
This should be equal to one of: \code{"location"},
\code{"status"},
one of the action identifiers with a "cost_" prefix,
one of the action identifiers with a "feasibility_" prefix,
or one of the action and feature identifiers with a "action_expectation_"
prefix.
For example, to show the cost of implementing action \code{"none"},
then the argument to data should be \code{"cost_none"}
Additionally, to show the action expectation for the action
\code{"none"} and the feature \code{"SPP1"}, then the argument should be
\code{"action_expectation_data_none&&&&&&&&&&&&SPP2"}.
Defaults to \code{"location"} such that the location of sites is shown
by displaying all sites with the same color.}

\item{\code{group}}{\code{character} group name. Defaults to \code{"sites"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} map.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_data}{}}}
\subsection{Method \code{get_solution_settings_data()}}{
Get data for rendering the solution settings widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_solution_settings_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_goals_settings_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_goals_settings_data}{}}}
\subsection{Method \code{get_goals_settings_data()}}{
Get data for rendering widget to specify goals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_goals_settings_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weights_settings_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_weights_settings_data}{}}}
\subsection{Method \code{get_weights_settings_data()}}{
Get data for rendering widget to specify weights.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_weights_settings_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_goal}{}}}
\subsection{Method \code{set_feature_goal()}}{
Set goal for a feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_feature_goal(feature_id, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_id}}{\code{character} value containing the feature identifier.}

\item{\code{value}}{\code{numeric} value containing the new goal.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_weight"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_weight}{}}}
\subsection{Method \code{set_feature_weight()}}{
Set weight for a feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_feature_weight(feature_id, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_id}}{\code{character} value containing the feature identifier.}

\item{\code{value}}{\code{numeric} value containing the new weight.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_site_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_site_data}{}}}
\subsection{Method \code{set_site_data()}}{
Set site data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_site_data(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{data.frame} containing new data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_data}{}}}
\subsection{Method \code{set_feature_data()}}{
Set feature data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_feature_data(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{data.frame} containing new data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feasibility_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_feasibility_data}{}}}
\subsection{Method \code{set_feasibility_data()}}{
Set feasibility data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_feasibility_data(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{data.frame} containing new data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_action_expectation_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_action_expectation_data}{}}}
\subsection{Method \code{set_action_expectation_data()}}{
Set action expectation data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_action_expectation_data(x, action_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{data.frame} containing new data.}

\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_site_data"></a>}}
\if{latex}{\out{\hypertarget{method-render_site_data}{}}}
\subsection{Method \code{render_site_data()}}{
Render site data using \pkg{rhandsontable}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_site_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_feature_data"></a>}}
\if{latex}{\out{\hypertarget{method-render_feature_data}{}}}
\subsection{Method \code{render_feature_data()}}{
Render feature data using \pkg{rhandsontable}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_feature_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_feasibility_data"></a>}}
\if{latex}{\out{\hypertarget{method-render_feasibility_data}{}}}
\subsection{Method \code{render_feasibility_data()}}{
Render feasibility data using \pkg{rhandsontable}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_feasibility_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_action_expectation_data"></a>}}
\if{latex}{\out{\hypertarget{method-render_action_expectation_data}{}}}
\subsection{Method \code{render_action_expectation_data()}}{
Render action expectation data using \pkg{rhandsontable}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_action_expectation_data(action_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write the data to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$write(workbook_path, geometry_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workbook_path}}{\code{character} file path.}

\item{\code{geometry_path}}{\code{character} file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
