% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Project.R
\name{Project}
\alias{Project}
\title{Project class}
\description{
Project class

Project class
}
\details{
Definition for the Project class.
}
\seealso{
\code{\link[=new_project]{new_project()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{site_ids}}{\code{character} vector}

\item{\code{site_descriptions}}{\code{character} vector}

\item{\code{feature_ids}}{\code{character} vector}

\item{\code{feature_html_ids}}{\code{character} vector}

\item{\code{feature_descriptions}}{\code{character} vector}

\item{\code{action_ids}}{\code{character} vector}

\item{\code{action_colors}}{\code{character} vector of colors for each action.}

\item{\code{action_descriptions}}{\code{character} vector}

\item{\code{site_data}}{\code{data.frame} object.}

\item{\code{feature_data}}{\code{data.frame} object.}

\item{\code{feasibility_data}}{\code{data.frame} object.}

\item{\code{site_geometry}}{\code{\link[sf:sf]{sf::st_sf()}} object.}

\item{\code{parameters}}{\code{list} of parameters.}

\item{\code{settings}}{\code{list} of \code{Parameter} objects.}

\item{\code{consequence_data}}{\code{list} of \code{data.frame} objects.}

\item{\code{site_cost_headers}}{\code{character} vector with names.}

\item{\code{site_status_header}}{\code{character} value.}

\item{\code{action_feasibility_headers}}{\code{character} vector with names.}

\item{\code{feature_goal_header}}{\code{character} value.}

\item{\code{feature_weight_header}}{\code{character} value.}

\item{\code{consequence_action_headers}}{\code{character} named vector.}

\item{\code{consequence_feature_headers}}{\code{character} named vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Project$new()}}
\item \href{#method-print}{\code{Project$print()}}
\item \href{#method-repr}{\code{Project$repr()}}
\item \href{#method-get_site_ids}{\code{Project$get_site_ids()}}
\item \href{#method-get_feature_ids}{\code{Project$get_feature_ids()}}
\item \href{#method-get_feature_ids_from_html_id}{\code{Project$get_feature_ids_from_html_id()}}
\item \href{#method-get_action_ids}{\code{Project$get_action_ids()}}
\item \href{#method-get_feature_goal}{\code{Project$get_feature_goal()}}
\item \href{#method-get_feature_weight}{\code{Project$get_feature_weight()}}
\item \href{#method-get_site_statuses}{\code{Project$get_site_statuses()}}
\item \href{#method-get_action_costs}{\code{Project$get_action_costs()}}
\item \href{#method-get_action_feasibility}{\code{Project$get_action_feasibility()}}
\item \href{#method-get_consequences_for_feature}{\code{Project$get_consequences_for_feature()}}
\item \href{#method-get_min_budget}{\code{Project$get_min_budget()}}
\item \href{#method-get_max_budget}{\code{Project$get_max_budget()}}
\item \href{#method-get_max_feature_consequence}{\code{Project$get_max_feature_consequence()}}
\item \href{#method-get_current_feature_consequence}{\code{Project$get_current_feature_consequence()}}
\item \href{#method-get_map_layers}{\code{Project$get_map_layers()}}
\item \href{#method-get_pu_data}{\code{Project$get_pu_data()}}
\item \href{#method-get_zone_data}{\code{Project$get_zone_data()}}
\item \href{#method-get_goal_data}{\code{Project$get_goal_data()}}
\item \href{#method-get_weight_data}{\code{Project$get_weight_data()}}
\item \href{#method-get_locked_data}{\code{Project$get_locked_data()}}
\item \href{#method-get_bbox}{\code{Project$get_bbox()}}
\item \href{#method-get_solution_settings_data}{\code{Project$get_solution_settings_data()}}
\item \href{#method-get_goals_settings_data}{\code{Project$get_goals_settings_data()}}
\item \href{#method-get_weights_settings_data}{\code{Project$get_weights_settings_data()}}
\item \href{#method-set_feature_goal}{\code{Project$set_feature_goal()}}
\item \href{#method-set_feature_weight}{\code{Project$set_feature_weight()}}
\item \href{#method-set_site_data}{\code{Project$set_site_data()}}
\item \href{#method-set_feature_data}{\code{Project$set_feature_data()}}
\item \href{#method-set_feasibility_data}{\code{Project$set_feasibility_data()}}
\item \href{#method-set_consequence_data}{\code{Project$set_consequence_data()}}
\item \href{#method-render_on_map}{\code{Project$render_on_map()}}
\item \href{#method-render_site_data}{\code{Project$render_site_data()}}
\item \href{#method-render_feature_data}{\code{Project$render_feature_data()}}
\item \href{#method-render_feasibility_data}{\code{Project$render_feasibility_data()}}
\item \href{#method-render_consequence_data}{\code{Project$render_consequence_data()}}
\item \href{#method-write}{\code{Project$write()}}
\item \href{#method-clone}{\code{Project$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Project object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$new(
  id,
  site_ids,
  site_descriptions,
  site_geometry,
  feature_ids,
  feature_descriptions,
  action_ids,
  action_descriptions,
  site_data,
  feature_data,
  feasibility_data,
  consequence_data,
  parameters
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} value.}

\item{\code{site_ids}}{\code{character} vector}

\item{\code{site_descriptions}}{\code{character} vector}

\item{\code{site_geometry}}{\code{\link[sf:sf]{sf::st_sf()}} object.}

\item{\code{feature_ids}}{\code{character} vector}

\item{\code{feature_descriptions}}{\code{character} vector}

\item{\code{action_ids}}{\code{character} vector}

\item{\code{action_descriptions}}{\code{character} vector}

\item{\code{site_data}}{\code{data.frame} object.}

\item{\code{feature_data}}{\code{data.frame} object.}

\item{\code{feasibility_data}}{\code{data.frame} object.}

\item{\code{consequence_data}}{\code{list} of \code{data.frame} objects.}

\item{\code{parameters}}{\code{list} of parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Project object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_site_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_site_ids}{}}}
\subsection{Method \code{get_site_ids()}}{
Get site identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_site_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_ids}{}}}
\subsection{Method \code{get_feature_ids()}}{
Get feature identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_feature_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_ids_from_html_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_ids_from_html_id}{}}}
\subsection{Method \code{get_feature_ids_from_html_id()}}{
Get feature HTML identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_feature_ids_from_html_id(feature_html_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_html_id}}{\code{character} vector of feature identifiers.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_ids}{}}}
\subsection{Method \code{get_action_ids()}}{
Get action identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_action_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_goal}{}}}
\subsection{Method \code{get_feature_goal()}}{
Get feature goal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_feature_goal(feature_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_id}}{\code{character} identifier for feature}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_weight"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_weight}{}}}
\subsection{Method \code{get_feature_weight()}}{
Get feature weight.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_feature_weight(feature_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_id}}{\code{character} identifier for feature}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_site_statuses"></a>}}
\if{latex}{\out{\hypertarget{method-get_site_statuses}{}}}
\subsection{Method \code{get_site_statuses()}}{
Get colors for actions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_site_statuses()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_costs"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_costs}{}}}
\subsection{Method \code{get_action_costs()}}{
Get cost data for implementing an action within each site.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_action_costs(action_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{numeric} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_feasibility"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_feasibility}{}}}
\subsection{Method \code{get_action_feasibility()}}{
Get feasibility data for implementing an action within each site.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_action_feasibility(action_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_consequences_for_feature"></a>}}
\if{latex}{\out{\hypertarget{method-get_consequences_for_feature}{}}}
\subsection{Method \code{get_consequences_for_feature()}}{
Get consequence data for features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_consequences_for_feature(action_id, feature_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}

\item{\code{feature_id}}{\code{character} identifier for feature.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{numeric} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_min_budget"></a>}}
\if{latex}{\out{\hypertarget{method-get_min_budget}{}}}
\subsection{Method \code{get_min_budget()}}{
Get the cost of the cheapest solution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_min_budget()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_max_budget"></a>}}
\if{latex}{\out{\hypertarget{method-get_max_budget}{}}}
\subsection{Method \code{get_max_budget()}}{
Get the cost of the most expensive solution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_max_budget()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_max_feature_consequence"></a>}}
\if{latex}{\out{\hypertarget{method-get_max_feature_consequence}{}}}
\subsection{Method \code{get_max_feature_consequence()}}{
Get the highest amount for each feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_max_feature_consequence()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} containing a \code{feature_id} and \code{amount} columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_current_feature_consequence"></a>}}
\if{latex}{\out{\hypertarget{method-get_current_feature_consequence}{}}}
\subsection{Method \code{get_current_feature_consequence()}}{
Get the current expected amount for each feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_current_feature_consequence()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} containing a \code{feature_id} and \code{amount} columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_map_layers"></a>}}
\if{latex}{\out{\hypertarget{method-get_map_layers}{}}}
\subsection{Method \code{get_map_layers()}}{
Get layer names for rendering data on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_map_layers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector of layer names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pu_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_pu_data}{}}}
\subsection{Method \code{get_pu_data()}}{
Get planning unit data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_pu_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_zone_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_zone_data}{}}}
\subsection{Method \code{get_zone_data()}}{
Get zone data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_zone_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_goal_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_goal_data}{}}}
\subsection{Method \code{get_goal_data()}}{
Get goal data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_goal_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weight_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_weight_data}{}}}
\subsection{Method \code{get_weight_data()}}{
Get weight data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_weight_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_locked_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_locked_data}{}}}
\subsection{Method \code{get_locked_data()}}{
Get locked data for optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_locked_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bbox"></a>}}
\if{latex}{\out{\hypertarget{method-get_bbox}{}}}
\subsection{Method \code{get_bbox()}}{
Get the bounding box.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_bbox(expand = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expand}}{\code{FALSE} should the bounding box be expanded by 10\%?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} object with \code{"xmin"}, \code{"xmax"}, \code{"ymin"}, and \code{"ymax"}
elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_settings_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_settings_data}{}}}
\subsection{Method \code{get_solution_settings_data()}}{
Get data for rendering the solution settings widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_solution_settings_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_goals_settings_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_goals_settings_data}{}}}
\subsection{Method \code{get_goals_settings_data()}}{
Get data for rendering widget to specify goals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_goals_settings_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weights_settings_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_weights_settings_data}{}}}
\subsection{Method \code{get_weights_settings_data()}}{
Get data for rendering widget to specify weights.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_weights_settings_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_goal"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_goal}{}}}
\subsection{Method \code{set_feature_goal()}}{
Set goal for a feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_feature_goal(feature_id, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_id}}{\code{character} value containing the feature identifier.}

\item{\code{value}}{\code{numeric} value containing the new goal.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_weight"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_weight}{}}}
\subsection{Method \code{set_feature_weight()}}{
Set weight for a feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_feature_weight(feature_id, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_id}}{\code{character} value containing the feature identifier.}

\item{\code{value}}{\code{numeric} value containing the new weight.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_site_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_site_data}{}}}
\subsection{Method \code{set_site_data()}}{
Set site data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_site_data(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{data.frame} containing new data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feature_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_feature_data}{}}}
\subsection{Method \code{set_feature_data()}}{
Set feature data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_feature_data(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{data.frame} containing new data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_feasibility_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_feasibility_data}{}}}
\subsection{Method \code{set_feasibility_data()}}{
Set feasibility data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_feasibility_data(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{data.frame} containing new data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_consequence_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_consequence_data}{}}}
\subsection{Method \code{set_consequence_data()}}{
Set consequence data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$set_consequence_data(x, action_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{data.frame} containing new data.}

\item{\code{action_id}}{\code{character} identifier for action.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_on_map(map, data = "location", group = "sites")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{data}}{\code{character} name of dataset to show.
Argument must be a valid layer name (see \code{self$get_map_layers()}).}

\item{\code{group}}{\code{character} group name. Defaults to \code{"sites"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} map.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_site_data"></a>}}
\if{latex}{\out{\hypertarget{method-render_site_data}{}}}
\subsection{Method \code{render_site_data()}}{
Render site data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_site_data(height = "100\%", width = "100\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}

\item{\code{width}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[rhandsontable:rhandsontable]{rhandsontable::rhandsontable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_feature_data"></a>}}
\if{latex}{\out{\hypertarget{method-render_feature_data}{}}}
\subsection{Method \code{render_feature_data()}}{
Render feature data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_feature_data(height = "100\%", width = "100\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}

\item{\code{width}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[rhandsontable:rhandsontable]{rhandsontable::rhandsontable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_feasibility_data"></a>}}
\if{latex}{\out{\hypertarget{method-render_feasibility_data}{}}}
\subsection{Method \code{render_feasibility_data()}}{
Render feasibility data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_feasibility_data(height = "100\%", width = "100\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{height}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}

\item{\code{width}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[rhandsontable:rhandsontable]{rhandsontable::rhandsontable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_consequence_data"></a>}}
\if{latex}{\out{\hypertarget{method-render_consequence_data}{}}}
\subsection{Method \code{render_consequence_data()}}{
Render consequence data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$render_consequence_data(action_id, height = "100\%", width = "100\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action_id}}{\code{character} identifier for action.}

\item{\code{height}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}

\item{\code{width}}{\code{character}/\code{numeric} CSS measurement value.
Default to \code{"100\%"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[rhandsontable:rhandsontable]{rhandsontable::rhandsontable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write the data to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$write(workbook_path, geometry_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workbook_path}}{\code{character} file path.}

\item{\code{geometry_path}}{\code{character} file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
