% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Solution.R
\name{Solution}
\alias{Solution}
\title{Solution class}
\description{
Solution class

Solution class
}
\details{
Definition for the Solution class.
}
\seealso{
\code{\link[=new_project]{new_project()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{project}}{\link{Project} object.}

\item{\code{settings}}{\code{list} of \link{Parameter} objects.}

\item{\code{summary_results}}{\code{data.frame} object.}

\item{\code{site_results}}{\code{data.frame} object.}

\item{\code{feature_results}}{\code{data.frame} object.}

\item{\code{solved}}{\code{logical} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Solution$new()}}
\item \href{#method-print}{\code{Solution$print()}}
\item \href{#method-repr}{\code{Solution$repr()}}
\item \href{#method-get_site_ids}{\code{Solution$get_site_ids()}}
\item \href{#method-get_feature_ids}{\code{Solution$get_feature_ids()}}
\item \href{#method-get_action_ids}{\code{Solution$get_action_ids()}}
\item \href{#method-get_bbox}{\code{Solution$get_bbox()}}
\item \href{#method-is_solved}{\code{Solution$is_solved()}}
\item \href{#method-render_on_map}{\code{Solution$render_on_map()}}
\item \href{#method-get_solution_results_data}{\code{Solution$get_solution_results_data()}}
\item \href{#method-render_summary_results}{\code{Solution$render_summary_results()}}
\item \href{#method-render_site_results}{\code{Solution$render_site_results()}}
\item \href{#method-render_feature_results}{\code{Solution$render_feature_results()}}
\item \href{#method-write}{\code{Solution$write()}}
\item \href{#method-clone}{\code{Solution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Solution object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$new(
  id,
  project,
  settings,
  summary_results,
  site_results,
  feature_results,
  solved
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} identifier.}

\item{\code{project}}{\code{character} identifier.}

\item{\code{settings}}{\code{list} of \link{Parameter} objects.}

\item{\code{summary_results}}{\code{data.frame} object.}

\item{\code{site_results}}{\code{data.frame} object.}

\item{\code{feature_results}}{\code{data.frame} object.}

\item{\code{solved}}{\code{logical} value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Solution object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repr"></a>}}
\if{latex}{\out{\hypertarget{method-repr}{}}}
\subsection{Method \code{repr()}}{
Generate a \code{character} summarizing the representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$repr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_site_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_site_ids}{}}}
\subsection{Method \code{get_site_ids()}}{
Get site identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_site_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_feature_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_feature_ids}{}}}
\subsection{Method \code{get_feature_ids()}}{
Get feature identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_feature_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_action_ids}{}}}
\subsection{Method \code{get_action_ids()}}{
Get action identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_action_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bbox"></a>}}
\if{latex}{\out{\hypertarget{method-get_bbox}{}}}
\subsection{Method \code{get_bbox()}}{
Get the bounding box.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_bbox(native = TRUE, expand = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{native}}{\code{logical} indicating if the bounding box should
be in (\code{TRUE}) the native coordinate reference system or (\code{FALSE})
re-projected to longitude/latitude?}

\item{\code{expand}}{\code{FALSE} should the bounding box be expanded by 10\%?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} object with \code{"xmin"}, \code{"xmax"}, \code{"ymin"}, and \code{"ymax"}
elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_solved"></a>}}
\if{latex}{\out{\hypertarget{method-is_solved}{}}}
\subsection{Method \code{is_solved()}}{
Is solved?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$is_solved()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_on_map"></a>}}
\if{latex}{\out{\hypertarget{method-render_on_map}{}}}
\subsection{Method \code{render_on_map()}}{
Render on map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_on_map(map, group = "sites")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map}}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.}

\item{\code{group}}{\code{character} group name. Defaults to \code{"sites"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}} map.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_solution_results_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_solution_results_data}{}}}
\subsection{Method \code{get_solution_results_data()}}{
Get data for rendering widget to display results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$get_solution_results_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_summary_results"></a>}}
\if{latex}{\out{\hypertarget{method-render_summary_results}{}}}
\subsection{Method \code{render_summary_results()}}{
Display summary results data using \pkg{rhandontable}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_summary_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_site_results"></a>}}
\if{latex}{\out{\hypertarget{method-render_site_results}{}}}
\subsection{Method \code{render_site_results()}}{
Display site results data using \pkg{rhandontable}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_site_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_feature_results"></a>}}
\if{latex}{\out{\hypertarget{method-render_feature_results}{}}}
\subsection{Method \code{render_feature_results()}}{
Display feature results data using \pkg{rhandontable}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$render_feature_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write the data to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$write(workbook_path, geometry_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workbook_path}}{\code{character} file path.}

\item{\code{geometry_path}}{\code{character} file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
