% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{whattodo}
\alias{whattodo-package}
\alias{whattodo}
\title{whattodo: Interactive management action prioritization application}
\description{
The What To Do application is a decision support tool to help prioritize
management actions for the Nature Conservancy of Canada. Data can be
uploaded using an Excel Spreadsheet and (optional) a shapefile delineating
the spatial location of sites. Prioritizations are generated using mixed
integer programming techniques. The performance of candidate prioritizations
can be examined using summary statistics and tables. Finally, data and
prioritizations can also be downloaded for subsequent analysis.
}
\examples{
\donttest{
if (interactive()) {
run_app()
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://NCC-CNC.github.io/whattodo/}
  \item Report bugs at \url{https://github.com/NCC-CNC/whattodo/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey O Hanson \email{jeffrey.hanson@uqconnect.edu.au} (\href{https://orcid.org/0000-0002-4716-6134}{ORCID})

Authors:
\itemize{
  \item Richard Schuster \email{richard.schuster@glel.carleton.ca} (\href{https://orcid.org/0000-0003-3191-7869}{ORCID})
}

}
